YARN-11664: Remove HDFS Binaries/Jars Dependency From Yarn
